Crate goat_cli

goat-cli is a command line interface to query the Genomes on a Tree Open API using an asynchronous tokio runtime.

I’m documenting the code here for others, and for future me.

Modules

count

Query the GoaT count API.

lookup

Query the GoaT lookup API.

progress

A module to produce a progress bar.

report

Query the GoaT record API.

search

Query the GoaT search API.

utils

Collection of utility functions used elsewhere.

Structs

GOAT_URL

The current GoaT URL.

TAXONOMY

The taxonomy that goat-cli uses.

UPPER_CLI_FILE_LIMIT

Upper limit for the number of entries in the file for CLI arg -f.

UPPER_CLI_SIZE_LIMIT

Upper limit for the CLI arg --size.

Enums

IndexType

The indexes we make searches over in GoaT.

TaxType

The type of result to return.

Constants

GOAT_API_VERSION 🔒

The current version of the GoaT API.

GOAT_URL_BASE 🔒

The base URL for GoaT.