Module goat_cli::utils::utils

Utility functions used across goat-cli.

Functions

did_you_mean

Shamelessly poached from the Nushell core code. Suggest the closest match to a string.

format_tsv_output

If multiple taxa are queried at once, headers will return for every new taxon. We can suppress this by storing the whole return as a string.

generate_unique_strings

Determine from the CLI matches how many URLs are needing to be generated, and return a vector of random character strings to use as unique identifiers.

get_rank_vector

Creates a vector of taxon ranks which will eventually form the headers of the taxon ranks in the returned TSV file.

levenshtein_distance 🔒

Compute the Levenshtein distance between two strings. Borrowed from here.

lines_from_file

Read NCBI taxon ID’s or binomial names of species, or higher order taxa from a file.

parse_comma_separated

Parse a comma separated string and return each of the elements stripped of whitespace in a vector.

pretty_print_usize

Thanks to this post on stack overflow. For error messages above cli query limit, print the usize prettily.

some_kind_of_uppercase_first_letter

Thanks to this post on stack overflow. Make a string uppercase on the first character.

switch_string_to_url_encoding

A function to replace certain combinations of characters as their URL encoded variations. Not entirely sure if this is necessary.